내부 클래스 추출 FragmentCnu, FragmentHome, FragmentSubject 룰 외부 클래스로 분리
