+++ app/src/main/res/menu/menu_main.xml
@@ -5,6 +5,6 @@
     <item
         android:id="@+id/action_settings"
         android:orderInCategory="100"
-        android:title="String"
+        android:title="Setting"
         app:showAsAction="never" />
 </menu>
